+++ CHANGELOG.md
@@ -8,7 +8,9 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 * New Features
    * Added the property `empty_partial` to the Constellation class
    * Added option to load Constellation for registered Instruments using lists
-     of platforms, names, tags, and/or inst_ids
+     of platforms, names, tags, and/or inst_ids, which are new attributes
+   * Added hidden Constellation method to determine unique attribute elements
+   * Extended Constellation unit tests
 * Documentation
    * Moved logo to 'docs\images'
    * Improved consistency of headers throughout documentation