+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [3.0.1] - 2021-07-27
+## [3.0.1] - 2021-07-28
 - New Features
   - Added a routine for loading CSV files into a pandas DataFrame from a list
     of filenames.